requirements.txt
streamlit
plotly
numpy
requests
pyyaml
